timesheet-app/pyproject.toml
[tool.poetry]
name = "timesheet-app"
version = "0.1.0"
description = ""
authors = ["Mark Francis <mark@markfrancis.io>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
lxml = "^4.7.1"
python-docx = "^0.8.11"
Flask-WTF = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
